fix: handle local deployment bundle edge cases
